+++ src/main/java/com/intellij/plugins/haxe/ide/lookup/HaxeSynteticLookupElements.java
@@ -1,18 +1,25 @@
 package com.intellij.plugins.haxe.ide.lookup;
 
-import com.intellij.plugins.haxe.lang.psi.fakes.HaxeFakeComponentBindMethod;
-import com.intellij.plugins.haxe.lang.psi.fakes.HaxeFakeComponentStringCode;
+import com.intellij.plugins.haxe.lang.psi.fakes.*;
 import icons.HaxeIcons;
 
 public class HaxeSynteticLookupElements {
 
     // all methods/functions have a bind method
-    public static HaxeLookupElement bind(HaxeFakeComponentBindMethod bind){
+    public static HaxeSynteticLookupElement bind(HaxeFakeNamedComponent bind){
         return new HaxeSynteticLookupElement(bind, "bind", "(...)","", HaxeIcons.Method, true);
     }
 
     // single character strings  have a "code" member/feature returning the character code (Int)
-    public static HaxeLookupElement code(HaxeFakeComponentStringCode stringCode) {
+    public static HaxeSynteticLookupElement code(HaxeFakeNamedComponent stringCode) {
         return new HaxeSynteticLookupElement(stringCode, "code", "","Int", HaxeIcons.Field, false);
     }
+    public static HaxeSynteticLookupElement trace(HaxeFakeNamedComponent stringCode) {
+        return new HaxeSynteticLookupElement(stringCode, "trace", "(v:Dynamic, ...customParams)","Void", HaxeIcons.Method, true);
+    }
+
+    public static HaxeSynteticLookupElement targetSpecificSyntax(HaxeFakeNamedComponent specificSyntax) {
+        return new HaxeSynteticLookupElement(specificSyntax, specificSyntax.getName(), "(code:String, args:Rest<Dynamic>)","Dynamic", HaxeIcons.Method, true);
+    }
+
 }