+++ src/main/java/com/intellij/plugins/haxe/lang/psi/fakes/HaxeFakeNamedComponent.java
@@ -1,11 +1,11 @@
 package com.intellij.plugins.haxe.lang.psi.fakes;
 
 import com.intellij.plugins.haxe.HaxeComponentType;
-import com.intellij.plugins.haxe.lang.psi.HaxeComponentName;
 import com.intellij.plugins.haxe.lang.psi.HaxeNamedComponent;
 import com.intellij.plugins.haxe.metadata.HaxeMetadataList;
 import com.intellij.plugins.haxe.metadata.psi.HaxeMeta;
 import com.intellij.plugins.haxe.metadata.psi.impl.HaxeMetadataTypeName;
+import com.intellij.plugins.haxe.model.HaxeMethodModel;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.tree.IElementType;
 import org.jetbrains.annotations.NotNull;
@@ -14,8 +14,6 @@ import org.jetbrains.annotations.Nullable;
 public abstract class HaxeFakeNamedComponent  extends HaxeFakePsiElement implements HaxeNamedComponent {
 
     private static final HaxeMetadataList EMPTY_META = new HaxeMetadataList();
-
-    public abstract HaxeComponentType componentType();
 
     @Override
     public @Nullable HaxeNamedComponent getTypeComponent() {
@@ -73,9 +71,17 @@ public abstract class HaxeFakeNamedComponent  extends HaxeFakePsiElement impleme
     }
 
     @Override
-    public String getDocs() {
+    public String getDocsText() {
         return "";
     }
+
+    @Override
+    public HaxeNamedComponent getDocsPsi() {
+        return null;
+    }
+
+    @Override
+    public abstract String getName();
 
     @Override
     public @NotNull HaxeMetadataList getMetadataList(@Nullable Class<? extends HaxeMeta> metadataType) {
@@ -87,4 +93,8 @@ public abstract class HaxeFakeNamedComponent  extends HaxeFakePsiElement impleme
         return false;
     }
 
+    @Nullable
+    public HaxeMethodModel getModel() {
+        return null;
+    }
 }